JpaAgreementRepository - save method always return null

+++ src/test/java/info/pionas/rental/infrastructure/persistence/jpa/agreement/JpaAgreementRepositoryIntegrationTest.java
@@ -36,7 +36,9 @@ class JpaAgreementRepositoryIntegrationTest {
 
     @AfterEach
     void deleteAgreements() {
-        springJpaAgreementTestRepository.deleteById(agreementId);
+        if (agreementId != null) {
+            springJpaAgreementTestRepository.deleteById(agreementId);
+        }
     }
 
     @Test
